invite

+++ src/services/business/TeamMemberServices.ts
@@ -10,6 +10,8 @@ import {
   IOrgMembership,
 } from "@/db/models/index.js";
 import ApiError from "@/utils/ApiError.js";
+import { invalidateCachesForUser } from "@/middleware/AddUserContext.js";
+import { invalid } from "joi";
 
 const SERVICE_NAME = "TeamMemberService";
 export interface ITeamMemberService {
@@ -39,6 +41,12 @@ class TeamMemberService implements ITeamMemberService {
   }
 
   create = async (orgId: string, data: Partial<ITeamMembership>) => {
+    const userId = data.userId?.toString();
+    if (!userId) {
+      throw new ApiError("User ID is required", 400);
+    }
+    invalidateCachesForUser(userId);
+
     const existing = await TeamMembership.findOne({
       teamId: data.teamId,
       userId: data.userId,
@@ -61,6 +69,7 @@ class TeamMemberService implements ITeamMemberService {
       ...data,
       orgId: new mongoose.Types.ObjectId(orgId),
     });
+
     return tm;
   };
 
@@ -127,6 +136,14 @@ class TeamMemberService implements ITeamMemberService {
   };
 
   delete = async (orgId: string, teamMemberId: string) => {
+    const tm = await TeamMembership.findOne({
+      _id: teamMemberId,
+      orgId: orgId,
+    });
+    if (!tm) {
+      throw new ApiError("Team member not found", 404);
+    }
+    invalidateCachesForUser(tm.userId.toString());
     await TeamMembership.findOneAndDelete({ _id: teamMemberId, orgId: orgId });
     return true;
   };
